+++ test/karma.config.js
@@ -27,7 +27,6 @@ module.exports = function(config) {
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
     preprocessors: {
-      // 'www/build/test/test.bundle.js': 'coverage',
       '**/*.ts': ['browserify']
     },